@@ -8,20 +8,28 @@
     "start": "next start"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@fortawesome/free-solid-svg-icons": "^5.15.2",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.2",
-    "@mdx-js/loader": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "@next/mdx": "^10.0.5",
-    "next": "10.0.5",
-    "prism-react-renderer": "^1.1.1",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "rehype-katex": "^4.0.0",
-    "remark-container": "^0.1.2",
-    "remark-math": "^3.0.0"
+    "@babel/core": "^7.0.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mdx-js/loader": "^2.2.1",
+    "@mdx-js/react": "^2.2.1",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.4",
+    "@mui/styles": "^5.11.2",
+    "@next/mdx": "^13.1.2",
+    "eslint": "^7.23.0",
+    "eslint-config-next": "^13.1.2",
+    "next": "^13.1.2",
+    "prism-react-renderer": "^1.3.5",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rehype-katex": "^6.0.2",
+    "remark-gfm": "^3.0.1",
+    "remark-math": "^5.1.1",
+    "typescript": "^4.9.4",
+    "webpack": "^5.0.0"
   }
 }
