@@ -1,6 +1,6 @@
 import Head from 'next/head';
 import Link from 'next/link';
-import Button from '@material-ui/core/Button';
+import Button from '@mui/material/Button';
 
 import config from '../app.config'; 
 
