@@ -2,7 +2,7 @@ import Head from 'next/head';
 import React, { useState } from 'react';
 
 import { Component } from "react";
-import { Button } from "@material-ui/core";
+import { Button } from "@mui/material";
 
 import Progress from "./Progress";
 
@@ -58,7 +58,7 @@ class GuidePane extends Component {
       savedStep = 0;
     }
 
-    this.setState({currentStep: savedStep});
+    // this.setState({currentStep: savedStep});
 
   }
 
